Implement GDAL type traits to convert a template to a GDAL data type.
Why? Syntactic sugar + ease of writing generic raster code. This module
allows you to write templated objects that can communicate their
basic type to GDAL. For example:

        template<typename T>
        class RowMajorMatrix<T> {
            static const GDALDataType kRasterType = gdaltypetraits::Convert<T>::value;

            void Write() {
               RasterIO(...args..., kGDALType);
            }
        }

    Voila! We can now seamlessly do things like this:

            RowMajorMatrix<float>::Write();
                  --- and ---
            RowMajorMatrix<int>::Write();

Like Boost, the internals of this module are wrapped in an `internal` namespace.
The only object outside the internal namespace maps a template to a gdal data
type:

    GDALDataType type = Convert<T>::value;

The structure within the internal namespace is a set of hierarchical templates.
Starting at the most generic level (ToGDAL<T>), the template type is specialized
until it reaches a GDALDataType. This hierarchy can handle floating point,
integral, boolean, and enumeration types.

Compile-time tests are included in a `test` namespace.